builder/generated/demo_worker/requirements.txt
temporalio
structlog
python-dotenv
